added a new hidden ip in form

+++ app.js
@@ -124,7 +124,10 @@ app.post("/", function(req, res){
 
 app.post("/delete", function(req, res){
   const checkedItemId =  req.body.checkbox;
-  Item.findByIdAndDelete(checkedItemId)
+  const listName = req.body.listName;
+
+  if(listName === "Today"){
+    Item.findByIdAndDelete(checkedItemId)
     .then(() =>{
       console.log("Sucessfulyy deleted item")
       res.redirect("/");
@@ -134,6 +137,9 @@ app.post("/delete", function(req, res){
       console.log(err)
     })
     
+  }
+
+ 
 })
 
 app.get("/work", function(req,res){
